h5io module puts h5 interaction fxns in separate module, away from plugin modules
